requirements.txt
pytubefix>=8.12.1
customtkinter>=5.2.2
pillow >= 11.1.0
requests >= 2.32.3
tkinter-tooltip >= 3.1.2
